pyproject.toml
[project]
name = "platspec-operator"
version = "0.1.0"
description = "Open-source Kubernetes operator implementing the Platform Specification processing layer"
authors = [
    { name = "Josh West", email = "josh@foundation.io" }
]
requires-python = ">=3.13,<3.14"
keywords = ["kubernetes", "operator", "kopf", "platform", "platformspec"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "kopf>=1.37.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "kubernetes>=28.1.0",
    "kcl-lib>=0.9.0,<0.12.0",
    "loguru>=0.7.0",
    "click>=8.1.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
platspec-operator = "platspec_operator.main:main"

[project.urls]
Homepage = "https://gitlab.com/platformspec/platspec-operator"
Repository = "https://gitlab.com/platformspec/platspec-operator"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/platspec_operator"]

[tool.hatch.version]
path = "src/platspec_operator/__init__.py"

[tool.ruff]
line-length = 88
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "integration: requires a live Kubernetes cluster with CRDs installed and operator running",
]

[tool.mypy]
python_version = "3.13"
packages = ["src/platspec_operator"]
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "kopf.*",
    "kcl_lib.*",
    "kubernetes.*",
]
ignore_missing_imports = true
